Adds new variant tests and configures dependency per variant.
Fixes #233

First stab at #60 but does not fix it since it is impossible to know whether or not the root is configured already.
