+++ fladle-plugin/src/test/java/com/osacky/flank/gradle/integration/TestFixtures.kt
@@ -17,3 +17,7 @@ fun TemporaryFolder.gradleRunner() =
     .withPluginClasspath()
     .forwardOutput()
     .withProjectDir(root)
+
+fun TemporaryFolder.writeEmptyServiceCredential() {
+  newFile("foo").writeText("{}")
+}
